+++ functions/types/modules.d.ts
@@ -10,22 +10,5 @@ declare module 'cookie-parser' {
   export default function cookieParser(secret?: string | string[], options?: unknown): RequestHandler
 }
 
-declare module 'graphql-got' {
-  interface GraphqlGotOptions {
-    query: string
-    headers?: Record<string, string>
-    variables?: Record<string, string>
-  }
-
-  interface GraphqlGotResponse<TBody = unknown> {
-    body: TBody
-  }
-
-  export default function graphqlGot<TBody = unknown>(
-    url: string,
-    options: GraphqlGotOptions
-  ): Promise<GraphqlGotResponse<TBody>>
-}
-
 declare module 'xml2js'
 declare module 'to-https'